query_executor: Fixed race between Query.Release() and speculative executions
As described in #88, there is a race condition when there is more than
one concurrent execution (by means of speculative execution) and
Query.Release() is called after one of them completes. Query.Release()
calls Query.reset(), which in turns zeroes the whole Query. Then, after
another execution completes, it tries to access metrics, but they are
already set to nil by the call to Release(), so a segfault is triggered.

The solution is quite simple: a ref counter is introduced into the Query.
It is obviously initially set to 1. Every execution fiber
(i.e. every execution goroutine) has the refcount atomically incremented
before it is started, and decrements the refcount on completion.
Query.Release() merely decrements the refcount. Query.reset() is called
by whichever goroutine that decrements the refcount to 0.